EDT-169: Include the Subversion revision in the assembly version. For example, the current version number is now 1.0.1244.240, not 1.0.1244. This is done only for a Release build; for a Debug build, revision number zero is used, to avoid problems for the developer and to make it easier to identify unofficial builds. For Release builds, the revision is extracted automatically from the last Subversion update; it therefore does not necessarily represent the head revision, but rather the revision on which the build is based.
The full version number is now displayed in the splash form, rather than just the build number that was previously shown. Instead of showing it in the form's title bar, it is shown in a label in the middle of the form. This allows the title bar and border to be hidden, as is normal form splash screens. In its other manifestation as the About box, the title bar is visible.