EDT-48: In the Participation Function box: * Add a vertical scroll bar to the Description text box. * The scroll bar on the list box was hidden off the edge of its parent control. This is fixed by widening the parent control.
These controls are also anchored properly to the right and bottom. This has the effect of widening the Description box in the Participation tab, but also making it a bit shorter; and, in other contexts where TextConstraintControl appears, it may fill the space to be longer than it used to be.