AWB-52: Merge TRUNK revisions 204 to 250 into BRANCHES/specialisation.
This does not compile. There is one error, which we will have to resolve:
ADL_Classes\ADL_Archetype.vb(1658) : error BC30456: 'constraint_model_factory' is not a member of 'openehr.adl_parser.interface.ARCHETYPE_PARSER'.

This utilises a finalised build of OceanInformatics.AdlParser.dll, built with EiffelStudio 6,1,7,1223 from BRANCHES/specialisation of ref_impl_eiffel, revision 407. Because this was not an official release build of the dll, the dll is kept in situ here. When eventually merging back into the Archetype Editor TRUNK, we will have to make sure that it uses an svn external as it currently does.