EDT-245: Enlarge the dialog that is shown when the user drags a Slot onto a Tree, to make sure that both items fit within the dialog's list box, even if the user has a font that is larger than usual.
Also, improve the behaviour. Previously it added an Element slot if the user simply closed the dialog; it now does not; and the user can now close it by typing the Esc key. It also used to close immediately when the user selected one of the items in the list; this was sensible behaviour if the user had done so via a mouse click, but it rendered the dialog almost unusable via the keyboard; it now closes by the user clicking with the mouse as before, or by the user typing the Enter key.