EDT-161: Dramatically reduce leak of Windows User Objects when switching to the Interface tab. Previously, when viewing openEHR-EHR-OBSERVATION.mst_colon.v1, 650 User Objects leaked each time you switched to the Interface tab, and when the total reached 10,000 the application crashed with an "error creating window handle" exception. It now leaks only 90 User Objects each time, making it far less likely to reach the limit.
The actual fix is at the top of Designer.BuildInterfaceTabPage(), which previously was clearing all controls from the Interface tab. Testing suggests that these controls were not being garbage-collected, because their Finalize() method was never called despite temporarily forcing garbage collection immediately after clearing the controls (i.e., the usual mantra: GC.Collect(); GC.WaitForPendingFinalizers(); GC.Collect()). Presumably something is keeping a reference to the controls contained immediately under the Interface tab; but despite extensive examination of the source code, including looking for event handlers, I could not find any such reference. The underlying cause of the leak therefore remains a mystery.

The fix is to call Dispose() for each of the Interface tab's dynamically-created controls. Calling Dispose() on a control also removes it from the tab's Controls list, so there is no need to clear the whole Controls list explicitly. Also, it is important to loop backwards through the Controls; looping forwards (e.g. with "for each") results in every second control being skipped.

There was an attempt to fix this problem a year ago (see EDT-24, revisions 128 and 129). A Dispose() method had been added to ArchetypeView; this is now removed because it was never called, and seems to have been completely misguided. A Dispose() method had also been added to ElementViewControl; this is now likewise removed, because although it made more sense, being an actual control container, the base class already does this correctly, so it is not needed. A lot of code that had been added to Designer and then commented out (presumably because it did not really work) has now been removed. (Note that EDT-24 therefore was completely unfixed, because all of the code that had been added in the course of the investigation was either commented out or not called.)

In the course of investigating this problem, I simplified a lot of the related code in order to make it easier to find the dangling reference to the controls. I'm committing these simplifications.