EDT-257: On the start-up window, don't allow text to be typed into the combo boxes.
This change was done in the Visual Studio forms designer. For some reason, the forms designer automatically changed the size of the group boxes, which meant that this form was now too big. It also had an adverse impact on the logic that hides the two group boxes at the bottom. The form has therefore been reduced in size, and the logic has been adjusted to compensate.