EDT-297: When saving an archetype via the Save dialog, if the user specifies an invalid archetype file name, then a valid file name based on the archetype id is used instead. For example, if the user omits the version part from the file name (e.g. "openEHR-EHR-CLUSTER.auscultation.adl") then it would be saved as "openEHR-EHR-CLUSTER.auscultation.v1.adl". Archetype Editor used to make this correction silently, however;  as well as being deceiving for the user, this was potentially dangerous, because the user might accidentally overwrite work that had not been backed up, possibly without realising that Archetype Editor had done this. Archetype Editor still does the correction, but it now displays a dialog explaining exactly what it is doing.
To make it less likely that the user will accidentally omit the version part, the Save dialog's file name text box includes the file extension (".adl" or ".xml"). For example, previously it showed "openEHR-EHR-CLUSTER.auscultation.v1", but it now shows "openEHR-EHR-CLUSTER.auscultation.v1.adl". This is mainly relevant if the user has the Windows Explorer option "Hide extensions for known file types" turned on, because then Windows automatically strips the file extension from what is visible in the Save dialog; therefore, prior to now, Windows would mistakenly think that the version ".v1" was the file extension, so the text box actually contained "openEHR-EHR-CLUSTER.auscultation" and the user was very likely to omit the version. The user is much less likely to make this mistake now, because if "Hide extensions for known file types" turned on the text box will now contain "openEHR-EHR-CLUSTER.auscultation.v1".