EDT-378: On beginning to drag a tree node that was already the selected node, the node went into edit mode. This bug is fixed by reverting a change made in revision 377 at the end of TreeStructure's DragDrop event handler, which for no apparent reason was putting the node into edit mode almost every time. The same line of code was also likely to cause a NullReferenceException; I saw this on one occasion while testing it.
There was a secondary bug. A stack overflow occurred if you dragged an anonymous slot onto a cluster that was already in edit mode. The stack overflow happens whenever the tree has two nodes in edit mode at the same time. Because this bug was a consequence of it going into edit mode during a drag, it has the same fix.

Having prevented nodes from going into edit mode during a drag operation, another problem became conspicuous. On dragging an anonymous slot, although it no longer went into edit mode, it still annoyingly changed into a named slot. This happened in a list too. This is fixed, in both the tree and the list, by changing it into a named slot on MouseUp rather than MouseDown. Furthermore, because it would be easy to cause this accidentally by clicking on an anonymous slot, a dialog now requests confirmation in the MouseUp event handler.