EDT-292: Loading embedded archetypes worked only if inclusion patterns were specified. "Include All", "Exclude All" and exclusion patterns are now taken into consideration too.
The absence of any constraint is treated the same as "Include All".

When inclusions and exclusions are both specified, the exclusions override the inclusions, except in the case of "Exclude All". I'm not sure whether this is the desired behaviour, but I guess it's better than completely ignoring them.

Also, fix a NullReferenceException if the user hits the OK button in the Choose dialog without selecting an archetype to embed.