EDT-458: Use the new ADL Parser DLL, which prevents the character sequence carriage return-carriage return-linefeed from being saved to an ADL file. This fixes a long-standing bug, where the multi-line text fields on the Description tab were saved to ADL with two newlines for each line that the user had entered.
In the multi-line text controls on the Description tab, prevent the character sequence carriage return-carriage return-linefeed. As well as causing additional carriage returns to be saved to the ADL file, the RTF on the Header tab was showing a space rather than a paragraph break because the RTF control does not know what to do with the illegal character sequence. This bug has lain dormant since being introduced in revision 117. It recently started causing an extra line to be inserted every time the user saved to ADL, because on opening an ADL file the process of inserting a carriage return in front of every linefeed was being called twice: once on loading and again on translating. The immediate cause of the bug therefore was probably the extra call added by revision 438. Most likely we can optimise this to ensure that the call is done only once, probably during translating. Be that as it may, for now it is safest to fix the root cause of the bug, i.e. the false assumption that there will not be any carriage returns yet, by replacing either each carriage return-linefeed pair or simple linefeed with the environment's newline string.