EDT-470 and EDT-471: When saving an archetype (either to ADL or to XML), if the archetype contained a slot or embedded archetype, if the regular expression specified the full archetype id (i.e. the new style of pattern, rather than the old style with only the concept and version parts), then the RM name and class were always prefixed to it (unless the RM type was abstract, e.g. Item). This caused the RM name and class to be duplicated again each time it was saved; e.g., after the third save it was 'openEHR-EHR-ITEM_TREE.openEHR-EHR-ITEM_TREE.openEHR-EHR-ITEM_TREE.referral(-[a-zA-Z0-9_]+)*\.v1'. This is fixed by using a different criterion for deciding whether to prefix the RM name and type; regardless of whether it's an abstract type, it now inserts the prefix if the RM name is not already present at the start of the pattern.
I decided to check only the RM name, not the RM class, for the benefit of EDT-471, where the existing archetype has a corrupt pattern, i.e. there is no backslash after the RM class (e.g. CLUSTER.foo instead of CLUSTER\.foo). Because the sole purpose of the prefix check is to decide whether the prefix is already there, in order to avoid inserting it again, checking the RM name is sufficient, and is more tolerant of errors that may appear in existing archetypes.

EDT-470 also noted that the dialog asking the user whether to load the embedded archetype was also duplicating the RM name and class. This is fixed by reorganising the logic, so that the question contains the actual file name that was found on disk. This means that this question, which previously was asked prior to searching the disk for archetypes, is now asked after the search has been done, and it is displayed only if exactly one matching archetype is found. As well as simplifying the source code, it streamlines the user interface in the cases where zero or multiple matching files exist, because instead of being asked whether to search, and then being asked to make another choice based on the search, the user is now just asked one question based on the result of the search.