Fix a bug when copying an RmEvent: one of the field variables was being copied to itself instead of to the other object.
As far as I can see, however, this Copy function is not called. It is certainly not called directly, but the call paths from the method that it overrides don't seem to lead to it either. In fact, there are a lot of Copy methods in many classes that have no callers. I would normally remove dead code like this, because it's a maintenance problem, but in this case I imagine they would be useful for implementing copy and paste functionality.