EDT-532: When opening an archetype with an invalid file name, the internal file name property was automatically corrected to match the archetype id, and a warning dialog was displayed so that the user knew that this was happening. This was good for most scenarios, but it was a nuisance to users who opened an archetype whose file name was being supplied in its shortened 8.3 form, for whatever reason. For example, this has been reported by one group of users who are opening archetypes from a mapped network drive by double-clicking the archetype in Template Designer. Regardless of whatever is causing the operating system to pass an 8.3 file name to Archetype Editor, there is no reason for Archetype Editor to be harassing these users about it whenever they open a file.
The fix is not to display this warning when opening an invalid file name. Instead, the dialog that was being displayed on opening the file appears when saving it. Rather than a warning, it is a question, asking the user whether to save under the valid file name derived from the archetype id. If the user chooses not to, then the file is not saved. This still prevents the user from saving an invalid file name, but lets the user back out of saving it. (Note that the "mismatched archetype id and file name" dialog is not appropriate in this scenario, because one of the two possible names -- the file name -- is invalid.)