EDT-93: Remove the Display tab's Save button. It saved the contents of the RTF control to a text file, rather than using the ADL or XML parser's save mechanism. This produced a file in Latin-1 encoding, not UTF-8. The button served no purpose other than to allow the RTF to be saved, which no one is likely to want to do anyway; if anyone does want to save the RTF, they can copy and paste it. Removing the button fixes the encoding bug. It also simplifies the user interface, because previously there were two "Save" buttons with subtly different behaviour.
While removing the button, I noticed that there was code for a context menu that allows searching of the text in the Display tab, but the context menu was not hooked up so it was never visible. I've hooked it up to the RTF control. Right-clicking the text now brings up the context menu, allowing you to search the ADL or XML. I've also put a "Find" button on the Display toolbar as an alternative to right-clicking.