EDT-599: Fix an IndexOutOfRangeException when opening an ADL file that contained a DV_COUNT with an unconstrained magnitude. The code was trying to access the children of the C_ATTRIBUTE with no children. The fix is to check whether the attribute has children before trying to access them.
There were dozens of other places where the source code was trying to do the same thing. These are all now protected in the same way.

Also, when displaying the ADL of an archetype containing a DV_COUNT with an unconstrained magnitude, there was a failing debug assertion. This is fixed by removing the assertion.