EDT-535: Fix exceptions and corruptions of the saved ADL and XML, caused by bad serialisation of a DV_CHOICE containing an INTERVAL<DV_DATE> or INTERVAL<DV_TIME>.
The bug was caused by revisions 523 and 524, which fixed EDT-361 by substituting "INTERVAL<DV_DATE_TIME>" with the appropriate RM type name. However, the fix for EDT-361 assumed that the INTERVAL<DV_DATE_TIME> would be the first child in the the DV_CHOICE, so it corrupted the first child's RM type name if something else happened to be in that position. The same error occurred for both ADL and XML.