EDT-608: In a specialised archetypes, the user was to rename a node defined in a parent without first specialising the node. This is always the wrong thing to do. Fix it such that, when the user attempts to rename the node, ask the user whether to specialise the node. If the user answers "No", then the renaming in not allowed. If the user answers "Yes", then the renaming is permitted.
Revert the changes from revision 608, because instead of asking the user whether to specialise the node, the user was still being shown the old warning message. This message is not longer relevant, so it has now been removed completely.

The fix has been done for tree, list, table and simple structures.

Note that the behaviour is imperfect in some respects. These issues might be dealt with by separate problem reports, if deemed worthwhile:
* In a tree, after the user answers "Yes", the node is no longer in edit mode. This is a bit unfriendly, but it is necessary in order to prevent a stack overflow.
* In a list, the node might become invisible. This seems to be a painting problem with the list control.
* In a table, the logic doesn't take effect until after the user tabs out of the grid cell. It might be nicer if it happened immediately.
* In a simple, the cursor jumps to the start of the text box. This could be annoying.