EDT-592, EDT-600 and EDT-615: Separate the XML Parser's CloneConstraintVisitor class into two classes:
(1) One is suitable for producing the canonical XML archetype model. It avoids tampering with the archetype model, and therefore stays close to the serialised ADL. Compared with what is generated by Archetype Editor 2.1, the only difference appears to be that the cardinality of clusters is now 1..* instead of 0..*. With that exception, then, the MD5 hash remains the same compared with Archetype Editor 2.1.

(2) The other is suitable for generating an OPT. It trims leading and trailing white space from term text values, removes empty term bindings, and sets existence to 0..1 rather than 1..1 where permitted by the reference model.

To achieve this, the changes that were done in revisions 591 and 592 to trim spaces, and the changes done in revision 600 to remove empty term bindings, have been moved from (1) to (2).