EDT-622: If an archetype's structure type is Single, then its element's occurrences must be 1. Prior to revision 593, this was enforced by making the occurrences control read-only. Revision 593 removed the line of code that achieved this, so allowing the user to select, say, an upper limit of 2, and on saving the archetype an error message was displayed.
Presumably the reason for deleting this line of code was that it would have been overriding other settings that had been achieved by the preceding lines of code, to fix other problems with slots and embedded archetypes.

The error is fixed by reinstating the deleted line of code, but placing it before rather than after those other changes that had been made in revision 593. This sets the occurrences control read-only for a Single structure type, as before, but without interfering with the fixes that were done in revision 593.