EDT-633: Allow the user to enter constraint bindings directly on the Definition tab, in the space below the Description text box, for any DV_TEXT that has the Terminology radio button selected.
The terminology, subset and release can be entered manually into three text boxes. Alternatively, the user can select the terminology and subset from OTS. A dialog is displayed for this purpose. The dialog also allows the user to browse the query subsets that have been defined for each terminology in OTS.

The format of the the URI that is saved to the constraint bindings section of the archetype is "terminology:terminology_id/release?subset=subset+name". The subset and release are optional. Pluses replace spaces in the subset name.

Only one constraint binding can be entered for each DV_TEXT. If the user wants more than one constraint binding, they can still do so under the Terminologies tab, on the Constraints sub-tab. The format of the URI that is created there has been changed to be the same as described above.

Note that, although this is working for basic scenarios, there are many pre-existing bugs associated with constraint bindings still remain. For example, if the terminology defined in the archetype does not match what is defined in the terminologies, then if the user switches to the Terminologies/Constraints tab and selects the affected constraint, an unhandled exception will appear repeatedly. Although this was an existing bug, if users start adding constraint bindings more than they had done in the past then they are more likely to bump into this problem.

Currently, if the user deletes a constraint binding on the Definition tab, it is not removed from the archetype. This should probably be considered a bug, but it is not clear what should be done since, having removed the terminology id, the program does not know which of the existing bindings was removed. This needs to be specified. In the meantime, the user can still remove the constraint binding via the Terminologies/Constraints tab.