EDT-424: Fix a problem on the Term Bindings tab, where terminologies that had been added via the Constraint Bindings tab were present but not visible in the drop-down list.
EDT-633: Allow the user to enter and view multiple constraint bindings directly on the Definition tab for any DV_TEXT that has the Terminology radio button selected. This is achieved by using the same grid that appears on the Terminologies/Constraints tab.

In the constraint bindings grid, display the subset instead of displaying the URI. This prevents the user from editing the URI directly, ensuring that it is formatted correctly.

Fix an unhandled exception that occurred repeatedly, if the terminology defined in the archetype does not match what is defined in the terminologies, when the user switched to the Terminologies/Constraints tab and selected the affected constraint. This is achieved by using a button rather than a data-bound combo box in the grid's Terminology column.

When serialising the constraint bindings URI, escape any characters that are not valid, e.g. the space character becomes "%20". (Note that the initial implementation of EDT-633 replaced the space with "+", but because there are many other invalid characters the same approach is now taken for all of them.)

When deserialising the URI, un-escape the encoded invalid characters. Fix a problem when extracting the subset parameter from the URI, where the whole parameter string was treated as the subset, and the subset was not extracted at all if it was not the first parameter.