EDT-633: The URI of a constraint bindings was not encoding the reserved characters defined in RFC 3986. For example, if the subset contains ':', which is a reserved character, this was stored directly in the URI (e.g., as "subset=openehr::foobar"). This was contrary to RFC 3986, which says, "If data for a URI component would conflict with a reserved character's purpose as a delimiter, then the conflicting data must be percent-encoded before the URI is formed." (http://tools.ietf.org/html/rfc3986#page-12)
This is fixed by encoding each component of the URI (the terminology, subset and release) individually, and then assembling the whole URI string. The reserved characters within each component are percent-encoded.

Deserialiasing the URI was likewise unescaping the whole URI. It now splits it into its components first, and then unescapes each component.