EDT-633: Archetypes created with old versions of Archetype Editor have constraint bindings that look like this: ["ac0001"] = <http://openEHR.org/p&t>
Archetype Editor was not displaying these old constraint bindings any more, because they are in the wrong form, i.e. they do not start with the scheme "terminology:". This is fixed by adding existing bindings to Archetype Editor's internal view, regardless of the URI.

The user can now see the URI directly in the constraint bindings grid. By hovering the mouse over the terminology cell, a tooltip showing the URI appears.