EDT-633: Fix a bug that prevented a constraint binding from being removed if a release had been specified for it. This problem occurred only for ADL archetypes, not for XML archetypes. It was caused by revision 352, which appended the release to the terminology id when adding or changing the constraint binding, but when attempting  to remove the constraint binding it ignored the release. The key therefore could not be found, and so it was not removed.
The fix is not to append the release to the terminology id. An alternative fix would have been to append it also when removing the binding. The approach of not appending it was chosen for consistency, because nothing but ADL constraint bindings did this: XML constraint bindings did not append the release (which is why they did not have this problem), nor did ADL or XML term bindings. (The latter do this in the code, but not on the terminology key, which is where the constraint bindings had it.)

Recent revisions had been working around various other problems that the appending of the release to the terminology id had caused, by including the revision in the key that was used for finding rows in the constraint bindings grid. This workaround has now been removed. The key now consists solely of the terminology id and the ac-code.

As a consequence of this, the first column of the grid now shows the true terminology key as it appears in the ADL. On opening an existing archetype that contains one of these old constraint bindings, the column may show, for example, "SNOWMED-CT(2003)" rather than "SNOMEDT-CT". The user can therefore see accurately what is going on now, and is free to modify or delete these old bindings if need be.