EDT-635: If a constraint binding was added and then removed, the terminology remained in the archetype. This was because adding a constraint binding added the terminology to the list of available terminologies, which had the side-effect of also adding an empty term binding to the same terminology (see EDT-597), and so the terminology remained when the constraint binding was removed because the term binding was still using it.
This problem only happened with ADL archetypes. It did not happen with XML archetypes. This was because, on adding a terminology, Archetype Editor was calling ARCHETYPE_ONTOLOGY.add_binding_terminology in the ADL Parser to add the terminology there. This routine adds the empty term binding. This looks like a bug in the routine, but as nothing except Archetype Editor appears to call it, it should probably be removed. This routine is redundant, because the terminology is added automatically whenever a term or constraint binding is added. Furthermore, Archetype Editor definitely should not have been calling it, because it is exported to NONE in the ARCHETYPE_ONTOLOGY class.

The fix is to remove the call to ARCHETYPE_ONTOLOGY.add_binding_terminology. This prevents the empty term binding, and so removing the constraint binding also removes the terminology, as expected, because nothing is using the terminology any more.

Note that this fix does not remove empty term bindings. It just prevents them from being created in the first place.