EDT-655: Fix a bug introduced in the last revision's refactoring of the TerminologySelectionForm, which was adding the terminology object rather than the queries list to the grid.
While fixing this, also remove an annoying side-effect that had been occurring while displaying the form in the past. When displaying the subsets of an ad-hoc terminology, the ad-hoc terminology was added to the list of terminologies. This was probably pointless, and probably misleading given that the ad-hoc terminology wasn't really available in the terminology server. It also tended to feel intermittent, since it only got added to the list after trying to display its subsets.