EDT-647: After selecting the option to change a structure type, the Save button was enabled, even if the user did not actually proceed with the change.
One reason why the Save button was becoming enabled was that the "Structure Type" combo box sometimes had the wrong type selected on loading the archetype. For example, if the structure was a list in an archetype where the default type was a tree, then the combo box showed "Tree". Changing in the combo box to "List" triggered the enabling of the Save button, despite the fact that the underlying structure actually was already a list. This is now fixed by ensuring that the combo box has the correct type selected whenever the "Change Structure" option is selected. (Apart from fixing the problem with the Save button, the fact that it was showing the wrong type was misleading to the user.)

Even after fixing the combo box, however, the Save button could still become enabled inappropriately if the user clicked on the combo box with the mouse. For some reason, clicking with the mouse triggers the selected index changed event, although hitting the <Esc> or <Enter> keys does not. To fix this, the logic in the event handler has now been reworked to ensure that the Save button is activated only if the structure type really is changing. (A complication to this is that it also needs to get enabled if the structure is actually an embedded archetype.)