EDT-647: Fix two problems with the last revision, which prevented the Save button from enabling after the user clicked the structure type without actually changing it: 1. The previous revision did not fix the problem for embedded archetypes. 2. It caused a new problem with embedded archetypes. If the user turned on the "Embedded" check box and then turned it off again, the display still appeared as if the check box was on. It did not show the structure controls.
Both problems are fixed by testing whether the user has altered the state of the check box. This involves comparing whether the check box is on against a flag. Previously this flag was updated at the very start of the check box's event handler. For the subsequent logic to be able to detect whether it has changed, we now update the flag towards the end of the event handler.

While making this change, I noticed that the event handler was popping open the structure type combo box if it did not have a selection yet. This is no longer necessary; as of revision 716, the combo box is always already initialised.