EDT-674: Fix a problem caused by revision 716, which initialised the Definition tab to a structure type appropriate for the kind of archetype (usually, a Tree). For an existing archetype, this caused an unused term to be added to the ontology section, representing the structure type, in addition to the term that is actually already in the existing archetype. This was happening because the constructor of the various structure user controls was always adding the new term to the ontology.
This was actually also the cause of an old bug that always existed prior to revision 716. Every time the user selected a different structure type, an extra, unused term was added to the ontology.

Both problems are fixed by changing the logic so that the decision whether to add a new term is no longer made from the structure control's constructor. Instead, it now happens later, after any existing structure has been copied into the control. This allows the structure control to check whether a term already exists for it; A new term is added only if there is not one already.

The existing term's text is now being updated correctly, therefore, instead of wrongly adding a new term. As a consequence, when the user selects a different structure type for an archetype that has translations, the question asking whether to update the translations of the term now pops up. Previously the question did not appear, because the text was not being updated. This question might surprise the user, so two changes have been made to it; both changes affect all scenarios where the question might appear, not just this particular scenario of changing the structure type:
1. The question is now displayed only if the new text is different from the old text. This prevents the question from appearing at all, in some cases.
2. When the question does appear, it would not have been clear to the user which term it was referring to. To help the user make sense of it, the question now states the at-code and text.

The way that the structure type is initialised has also been changed from the way that it was done in revision 716. Previously, the initialisation was being done while the structure type combo box was being populated. I had at first suspected that this might have been the cause of the bug, on the supposition that by doing it so early in the lifetime of the control, the combo box selection would be changing more frequently than necessary. I initially tried to fix it, then, by moving the initialisation to occur during loading of the tab page. Although this didn't work, it does seem more logical. Care has to be taken, however, to set the loading flag, to prevent the Save button from being enabled while initialising the combo box.

While testing these changes, I found that they caused the Save button to become enabled on loading an instruction archetype with an embedded activity. The root cause of this new problem turned out to be that the various structure user controls were enabling the Save button in their constructors, for no apparent reason. As these constructors are called only from the combo box's selection changed event handler, and as this event handler already looks after the enabling of the Save button correctly, the various constructors no longer do this. This prevents the Save button from being enabled inappropriately on loading the archetype.