EDT-679: Fix an index out of range exception that occurred when saving an ADMIN_ENTRY that had no items in the definition section yet.
This is similar to the fix in revision 706. This problem was not occurring in revision 735. This problem is likely to have been caused by revision 763. Note that the ADL generated by revision 735 contained an extra child node compared with older revisions (e.g. revision 583). With this current fix, the ADL no longer contains the extra child node, and the definition section is exactly the same as revision 583.