EDT-691: Fix a bug caused by revision 768. Specialising a quantity element cleared the property and corrupted it in the saved ADL.
The bug was actually not caused by that particular revision, but by the fact that the quantity constraint class's Copy method did not copy all of the field variables to the new quantity constraint. Fixing this fixes the problem.

Inspection of the Copy method of the other constraint classes showed that most of them also failed to copy some of the field variables. I didn't try to discover whether these were causing similar bugs during specialisation, but no doubt they would have been.