EDT-704: Fix an exception that occurred while serialising an ELEMENT archetype to XML if the element had a name constraint. The problem was caused by the name constraint being duplicated.
The same bug had already been fixed in revision 529 for EDT-509. I've taken the same approach to fixing it for XML. In theory, however, the bug could occur for any type of archetype that has a name constraint on at0000, so I'm not convinced that this is the best fix. I think it might be better to revert this revision as well as revision 529, and to remove the code in MakeParseTree() that is adding the name constraint to the root node. But given that Archetype Editor currently can't, as far as I'm aware, add a name constraint to the root node except for ELEMENT archetypes, the current fix is quicker to test and less risky.