EDT-640: Fix problems that occurred adding a second language to an archetype on the Terminology tab, under "Languages & Terminologies", and then saving as XML. First of all, there were some null reference exceptions because the second language has not been added to the list of translations yet in the description section.
Once these were fixed, allowing the archetype to be saved as XML, there was still a problem that the XML contained the new language in the details of the description section. The XML failed schema validation because the language's details were incomplete. The fix was to prevent the new language from being added to the details, to be consistent with the ADL.

There was still another problem. On opening the XML archetype, the new language was missing. This was happening because the XML parser calculating the available languages incorrectly, looping through all of the translations, and because the new language isn't there it wasn't loading it. The problem is fixed by doing what the ADL parser does, i.e., instead of looping through the translations, it now loops through all of the term definitions.

Finally, after all of the problems loading the XML archetype were fixed, there was a null reference exception if you then tried to display it as ADL. This was similar to the initial exceptions that were happening when saving as XML, and the fix is similar.