EDT-653 and EDT-693: On the Definition tab of Single, List, Tree and Table archetypes, hide the structure panel. Showing it was pointless, because the structure type is fixed for these archetypes.
Hiding it also removes the ability to embed another archetype, which made no sense from a practical point of view. From a technical point of view, it was impossible to generate ADL for this, because there's no parent node for the embedded archetype.

While investigating this, I cleaned up the code for making the parse tree of an XML archetype, to be more consistent with the ADL code. This mainly affects OBSERVATION and ADMIN_ENTRY XML archetypes.

Also, because CLUSTER and ELEMENT archetypes cannot contain an embedded archetype, I've removed some code from the structure tab page control that was supporting this.