EDT-323: In specialised archetypes, prevent editing of a node unless the user has already specialised it. This is achieved by disabling the node's details panel over on the right.
To make it clear to the user why the details are disabled, a "Specialise" button is displayed above them. This duplicates the behaviour of the existing "Specialise" option in the node's context menu.

In the structure tab control, the "Structure" combo box and the "Embedded" check box are now always read-only for specialised archetypes.

There was no "Specialise" option in the context menu of the section and composition content tree views. This has now been implemented, so that the "Specialise" button works the same there as on the other structure controls.