EDT-749: Allow Archetype Editor to be run in a batch-processing mode from the command line, so that multiple ADL or XML archetypes can be opened one by one, writing each archetype as ADL or XML or both.
The command line options are:
ArchetypeEditor.exe [languagecode] [archetypefiles] [-exportadl] [-exportxml] [exportdirectory]

Previously, the first two options were already accepted, but the [archetypefiles] option could be only one archetype file. For batch processing, it now allows wildcard patterns (e.g., "knowledge\*adl" or "knowledge\*xml").

Both export flags (i.e., -exportadl and -exportxml) can be specified, so that ADL and XML will both be generated.

The [exportdirectory] option is optional. If it is not specified, then each archetype is written out to the same directory as the source archetype file. (This means that the source archetype is overwritten, if it is output in the same format.)

For example, if you have a directory of XML archetypes, you can generate a directory containing ADL archetypes with a command similar to this:
ArchetypeEditor.exe C:\knowledge\*.xml -exportadl C:\exported